+++ libs/base/src/code/error/mod.rs
@@ -3,7 +3,8 @@
 #![doc = crate::_DOC_CODE_ERROR!()]
 //
 
-mod definitions; // modular errors
+mod define_error; // define_error!
+mod errors; // general errors
 mod reexports;
 
 crate::items! { // structural access: _mods, _all
@@ -11,7 +12,7 @@ crate::items! { // structural access: _mods, _all
     pub use _mods::*;
 
     mod _mods {
-        pub use super::{definitions::*, reexports::*};
+        pub use super::{define_error::*, errors::*, reexports::*};
     }
     pub(super) mod _all {
         #[doc(inline)]