+++ libs/base/src/code/util/mod.rs
@@ -13,7 +13,6 @@ mod reexports; // re-exported macros from devela_base_macros
 
 mod cdbg; // cdbg!
 mod cfg_if; // cfg_if!
-mod define_error; // define_error!
 mod deprecate; // deprecate_feature!
 mod include; // include_from!, mod_from!
 mod items; // items!, sf!
@@ -32,7 +31,7 @@ crate::items! { // structural access: _mods, _workspace_private, _all
 
     mod _mods {
         pub use super::{
-            cdbg::*, cfg_if::*, define_error::*, deprecate::*, include::*, items::*, is::*,
+            cdbg::*, cfg_if::*, deprecate::*, include::*, items::*, is::*,
             paste::*, r#const::*, reexports::*,
         };
     }